challenges/web/jwthe/Docker/src/requirements.txt
cryptography
PyJWT==2.6.0